minor review style fixes
<==== Mercurial data ====>
Branch: dev_3.0

+++ unit_tests/nx_network_ut/src/simple_socket_test_helper.h
@@ -14,7 +14,7 @@ namespace /* anonimous */ {
 
 const SocketAddress kServerAddress("127.0.0.1:12345");
 const QByteArray kTestMessage("Ping");
-const size_t kClientCount(3);
+const int kClientCount(3);
 const std::chrono::milliseconds kTestTimeout(500);
 
 template<typename ServerSocketMaker, typename ClientSocketMaker>
